+++ core/src/main/java/com/graphhopper/routing/util/EncodingManager.java
@@ -167,9 +167,9 @@ public class EncodingManager
      * <p/>
      * @return the encoded flags
      */
-    public int handleWayTags( int includeWay, OSMWay way )
+    public long handleWayTags( int includeWay, OSMWay way )
     {
-        int flags = 0;
+        long flags = 0;
         for (int i = 0; i < encoderCount; i++)
         {
             flags |= encoders.get(i).handleWayTags(includeWay, way);
